pyproject.toml
[tool.poetry]
name = "ragie-webhook-ref-app"
version = "0.1.0"
description = ""
authors = ["Matt Kauffman <matt@ragie.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = "^0.31.1"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
